refactor(host-infra-beads): split Beads adapter monolith into focused modules
Extract command execution, initialization guards, parsing and metadata concerns, and task store orchestration into dedicated files to reduce coupling and lower regression risk when extending Beads integration.
