+++ apps/desktop/src-tauri/crates/host-infra-beads/src/constants.rs
@@ -1,0 +1,3 @@
+pub(crate) const DEFAULT_METADATA_NAMESPACE: &str = "openducktor";
+pub(crate) const CUSTOM_STATUS_VALUES: &str = "spec_ready,ready_for_dev,ai_review,human_review";
+
